+++ BookStore/src/main/java/io/jexxa/tutorials/bookstore/domainservice/IBookRepository.java
@@ -1,0 +1,24 @@
+package io.jexxa.tutorials.bookstore.domainservice;
+
+import io.jexxa.addend.applicationcore.Repository;
+import io.jexxa.tutorials.bookstore.domain.aggregate.Book;
+import io.jexxa.tutorials.bookstore.domain.valueobject.ISBN13;
+
+import java.util.List;
+import java.util.Optional;
+
+@Repository
+public interface IBookRepository
+{
+    void add(Book book);
+
+    Book get(ISBN13 isbn13);
+
+    boolean isRegistered(ISBN13 isbn13);
+
+    Optional<Book> search(ISBN13 isbn13);
+
+    void update(Book book);
+
+    List<Book> getAll();
+}
