+++ BookStore/src/main/java/io/jexxa/tutorials/bookstore/domainservice/IDomainEventPublisher.java
@@ -1,0 +1,9 @@
+package io.jexxa.tutorials.bookstore.domainservice;
+
+import io.jexxa.addend.applicationcore.InfrastructureService;
+
+@InfrastructureService
+public interface IDomainEventPublisher
+{
+    void publish(Object domainEvent);
+}
