--- BookStoreJ16/pom.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <artifactId>jexxatutorials</artifactId>
-    <groupId>io.jexxa.tutorials</groupId>
-    <version>1.0.16-SNAPSHOT</version>
-  </parent>
-
-  <name>BookStoreJ16</name>
-  <artifactId>bookstorej16</artifactId>
-  <groupId>io.jexxa.tutorials.bookstorej16</groupId>
-
-  <properties>
-    <main.arguments>-jdbc -jms</main.arguments>
-    <main.http.port>7504</main.http.port>
-    <main.https.port>8084</main.https.port>
-
-    <maven.javadoc.skip>true</maven.javadoc.skip>
-    <maven.compiler.release>16</maven.compiler.release>
-  </properties>
-
-</project>
