--- BookStoreJ16/src/main/java/io/jexxa/tutorials/bookstorej16/domain/businessexception/BookNotInStockException.java
@@ -1,12 +1,0 @@
-package io.jexxa.tutorials.bookstorej16.domain.businessexception;
-
-import io.jexxa.addend.applicationcore.BusinessException;
-
-/**
- * Is thrown in case we try to sell a book that is currently not in stock
- */
-@BusinessException
-public class BookNotInStockException extends Exception
-{
-    private static final long serialVersionUID = 1L;
-}
