--- BookStoreJ16/src/main/java/io/jexxa/tutorials/bookstorej16/domain/domainevent/BookSoldOut.java
@@ -1,13 +1,0 @@
-package io.jexxa.tutorials.bookstorej16.domain.domainevent;
-
-import io.jexxa.addend.applicationcore.DomainEvent;
-import io.jexxa.tutorials.bookstorej16.domain.valueobject.ISBN13;
-
-@DomainEvent
-public record BookSoldOut(ISBN13 isbn13)
-{
-    public static BookSoldOut bookSoldOut(ISBN13 isbn13)
-    {
-        return new BookSoldOut(isbn13);
-    }
-}
