--- BookStoreJ16/src/main/java/io/jexxa/tutorials/bookstorej16/domainservice/IDomainEventPublisher.java
@@ -1,9 +1,0 @@
-package io.jexxa.tutorials.bookstorej16.domainservice;
-
-import io.jexxa.addend.applicationcore.InfrastructureService;
-
-@InfrastructureService
-public interface IDomainEventPublisher
-{
-    void publish(Object domainEvent);
-}
